% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignElevation.R
\name{assignElevation}
\alias{assignElevation}
\title{Assigning elevation data}
\usage{
assignElevation(data = NULL, path.data = NULL, rd.frmt = "readRDS",
  elevations.db = NULL, round.coord = 1, wrt.frmt = "saveRDS",
  save.assigned.in = NULL, save.unassigned.in = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{path.data}{Vector of characters. Path to the input file.}

\item{rd.frmt}{Vector of characters. The file format to read. 
By default it will be read as a R object using the
\code{'readRDS'} argument, but it can be read as a plain text file, using the
\code{'readTXT'} argument. See details.}

\item{elevations.db}{Data.Frame object. The elevations database, 
tree columns (decimalLatitude, decimalLongitude, elevation).}

\item{round.coord}{Numeric vector. Decimal to round the coordinate to
assign the elevation. This value must be same of the coordinate 
resolution of database assigned as argument on \code{elevation.db} parameter. 
Example: the resolution of the coordinates in the databases assigned on 
\code{elevation.db} is one decimal, then \code{round.coord} must be equal to 1 (round.coord=1).}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using the
\code{'saveRDS'} argument, but it can be saved as plain text using the
\code{'saveTXT'} argument. See details.}

\item{save.assigned.in}{Vector of characters. Path to
write the output file with only the coordinates whose elevation was assigned. 
Output file will be called <alt.assigned>.}

\item{save.unassigned.in}{Vector of characters. Path to 
write the output file with only the coordinates whose elevation was not assigned. 
Output file will be called <alt.unassigned>. See details.}
}
\value{
A table data.frame class with coordinates and their respective elevation. 
A table with coordinates whose elevations were not assigned. A table of information.
}
\description{
Given a geographical coordinate (decimalLongitude/ decimalLatitude)
the elevation data is assigned from an elevation database (Headers: decimalLongitude/
decimalLatitude/ elevation).
}
\details{
For more details about the formats to read and/or write, see the
\code{\link{readAndWrite}} function.

The coordinates whose elevations were not assigned could be checked with
\code{\link{elevFromGg}}. This function can download data directly from 
Google Maps Elevation API [1]

The headers of the input file must follow the Darwin Core standard [2]. 
The user can see the guide using data('ID_DarwinCore) command.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] The Google Maps Elevation API.  Available online at \url{https://developers.google.com/maps/documentation/elevation/intro}

[2] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{elevFromGg}}

\code{\link{readAndWrite}}
}

