% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chcbiog.R
\docType{data}
\name{ChcBiog}
\alias{ChcBiog}
\title{Biogeographic Choco  (Sensu: Hernandez-Camacho et al. 1992) [1]}
\format{Format class \code{'SpatialPolygonsDataFrame'} [package \code{'sp'}] with 5 slots: data, polygons, plotOrder, bbox, proj4string, keywords.}
\usage{
data(ChcBiog)
}
\description{
Polygon of  the Biogeographic Choco (Sensu: Hernandez-Camacho et al. 1992)[1] using as example in R-Alarcon and Miranda-Esquivel (In prep)[2].
The \code{'ChcBiog'} object uses the WGS84 datum.
}
\examples{
\donttest{
data(ChcBiog)
plot(ChcBiog)}
}
\references{
[1] Hernandez-Camacho, J. et al. 1992. Unidades biogeograficas de Colombia, In: Halffter, G. (ed), La Diversidad Biologica de Iberoamerica I. Acta Zoologica Mexicana. Instituto de Ecologia, A.C., pp. 105-151
[2] R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to cleaning geographical data from electronic biodatabases.
}
\keyword{datasets}

