% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCoord.R
\name{checkCoord}
\alias{checkCoord}
\title{Check coordinates}
\usage{
checkCoord(data = NULL, path = NULL, rd.frmt = "readRDS",
  wrt.frmt = "saveRDS", save.right.in = NULL, save.wrong.in = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{path}{Vector of characters. Path to the input file.}

\item{rd.frmt}{Vector of characters. The file format to read.
By default it will be read  as a  R object using
\code{'readRDS'} argument, but it can be read as plain text using
\code{'readTXT'} argument. See details.}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using
\code{'saveRDS'} argument, but it can be saved as plain text using
\code{'saveTXT'} argument. See details.}

\item{save.right.in}{Vector of characters. Path to the output
file for each species with coordinates in the correct format.}

\item{save.wrong.in}{Vector of characters. Path to the output
file for each species with coordinates in the wrong format.}
}
\value{
For each species, if this is the case, two output files
data.frame class: The first one has the correct coordinates for the species and
the second one has the wrong coordinates for the same species.
}
\description{
Separate coordinates with a format different to decimal degrees
and check that the coordinates in this format are into range (latitude: -90, 90 degrees and
longitude: -180, 180 degrees)
}
\details{
For more details about the formats to read and/or write, see
\code{\link{readAndWrite}} function.

The headers of the input file must follow the Darwin Core standard [1].
The user can see the guide using data('ID_DarwinCore) command.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR. (submitted.) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

