% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordToISO.R
\name{coordToISO}
\alias{coordToISO}
\title{From coordinate in decimal degree format to International Organization for Standardization (ISO 3166-1 alpha-2/alpha-3)}
\usage{
coordToISO(coordinates.tab = NULL, iso = 2)
}
\arguments{
\item{coordinates.tab}{Data Frame class. A data frame with at least two columns
called \code{decimalLatitude} and \code{'decimalLongitude'}.}

\item{iso}{numeric (2 or 3). To define the code whether ISO-2 (iso=2, default value) or ISO-3 (iso=3).}
}
\value{
A table with three columns (decimalLatitude, decimalLongitude, countryCode)
}
\description{
Give a coordinate in decimal degree format to assign the
corresponding International Organization for Standardization (ISO 3166-1 alpha-2/alpha-3) [1].
}
\details{
If the coordinate is outside the continent, this can not be assigned and
will be assigned as NaN. If you want to see the oficial ISO codes, you could use the data(ISO2) or data(ISO3) comands.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1]ISO 2015. International standard for country codes and codes for their subdivisions. Country codes - iso 3166.  International Organization for Standardization. Available online at \url{http://www.iso.org/iso/country_codes}
}
\seealso{
\code{\link{coordToCountry}}
}

