% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryNames.R
\docType{data}
\name{countryNames}
\alias{countryNames}
\title{Countries Names}
\format{A vector class. Factor with 246 levels (Countries Names).}
\usage{
data(countryNames)
}
\description{
List of Countries Names in English.
}
\examples{
 \donttest{data(countryNames)}

}
\references{
R-Alarcon V. and Miranda-Esquivel DR. (submitted) geocleaMT: An R
package to cleaning geographical data from electronic biodatabases.
}
\keyword{datasets}

