% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryToISO.R
\name{countryToISO}
\alias{countryToISO}
\title{From country name to International Organization for Standardization (ISO 3166-1 alpha-2 / alpha-3)}
\usage{
countryToISO(country = NULL, iso = 2)
}
\arguments{
\item{country}{Data.Frame class. a vector with countries names in English. See
details.}

\item{iso}{numeric (2 or 3). To define the code whether ISO-2 (iso=2, default value) or ISO-3 (iso=3).}
}
\value{
A data frame object with the ISO 3166-1 alpha-2/alpha-3 assigned to each country.
}
\description{
From country name in English to International Organization for Standardization
(ISO 3166-1 alpha-2 / alpha-3) [1].
}
\details{
To see the valid names in English the user can use the command
data(countryNames).
}
\note{
See:
R-Alarcon V, Miranda-Esquivel DR.(submitted) geocleaMT: An R package to cleaning
geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1]ISO 2015. International standard for country codes and codes for their subdivisions. Country codes - iso 3166.  International Organization for Standardization. Available online at \url{http://www.iso.org/iso/country_codes}.
}
\seealso{
\code{\link{countryToCentroid}}
}

