% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutRange.R
\name{cutRange}
\alias{cutRange}
\title{Species on a specific elevation range}
\usage{
cutRange(data = NULL, path = NULL, rd.frmt = "readRDS", range.from = 0,
  range.to = 1000, wrt.frmt = "saveRDS", save.inside.in = NULL,
  save.outside.in = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{path}{Vector of characters. Path to the input file.}

\item{rd.frmt}{Vector of characters. File format to read. 
By default it will be read  as a  R object using the
\code{'readRDS'} argument, but it can be read as plain text using the
\code{'readTXT'} argument. See details.}

\item{range.from}{Numeric vector. Lower bound of the range of the distribution}

\item{range.to}{Numeric vector. Upper bound of the range of the distribution.}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using the
\code{'saveRDS'} argument, but it can be saved as plain text using the
\code{'saveTXT'} argument. See details.}

\item{save.inside.in}{Vector of characters. Path to the output 
file with the species whose occurrences are on the range 
assigned in \code{range.from} to \code{range.to} parameters. Output file 
will be called <inside.range>.}

\item{save.outside.in}{Vector of characters. Path to the output 
file with the species whose occurences are not on the range 
assigned in \code{range.from} and \code{range.to} parameters. Output file 
will be called <outside.range>.}
}
\value{
A table data.frame class with only the species whose occurences is on range 
assigned in \code{range.from} and \code{range.to} parameters, and a vector with a table with descriptive quantities.
}
\description{
Get species distributed on a specific range of altitude.
}
\details{
For more details about the formats to read and/or write, see 
\code{\link{readAndWrite}} function.

The headers of the input file must follow the Darwin Core standard [1]. 
The user can see the guide using data('ID_DarwinCore) command.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

