% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltPointOrSp.R
\name{delPointsOrSp}
\alias{delPointsOrSp}
\title{Delete some/all records of a given species}
\usage{
delPointsOrSp(data = NULL, rd.frmt = "readRDS", path = NULL,
  west = -120, east = -65, south = -30, north = 30, plot.distrib = T,
  wrt.frmt = "saveRDS", save.file = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{rd.frmt}{Vector of characters. The file format to read. 
By default it will be read  as a  R object using the
\code{'readRDS'} argument, but it can be read as plain text using the
\code{'readTXT'} argument. See details.}

\item{path}{Vector of characters. Path to the input file.}

\item{west}{Numeric vector. Western coordinate in decimal longitude format.}

\item{east}{Numeric vector. Eastern coordinate in decimal longitude format.}

\item{south}{Numeric vector. Southern coordinate in decimal latitude format.}

\item{north}{Numeric vector. Northern coordinate in decimal latitude format.}

\item{plot.distrib}{Logical vector. If \code{'TRUE'}, the distribution of the species will 
be plotted in red on a world map.}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using 
\code{'saveRDS'} argument, but it can be saved as plain text using 
\code{'saveTXT'} argument. See details.}

\item{save.file}{Vector of characters. Path to the output 
file for each  species.}
}
\value{
A file for each species as data frame class.
}
\description{
The area of interest is framed as the four coordinates (North, 
South, West and East), if the species has some points outside of this area the user
will be asked to decide whether delete the point or delete the  species.
}
\details{
The headers of the input file must follow the Darwin Core standard [1]. 
The user can see the guide using data('ID_DarwinCore) command.
For  more details about the formats to read and/or write, see 
\code{\link{readAndWrite}} function.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

