% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevFromGg.R
\name{elevFromGg}
\alias{elevFromGg}
\title{Assign elevation to a coordinates list}
\usage{
elevFromGg(data = NULL, rd.frmt = "readRDS", path = NULL,
  API.Key = NULL, starts.in = 1, round.coord = 4, save.name = "data",
  wrt.frmt = "writeRDS", save.assigned.in = NULL,
  save.unassigned.in = NULL, save.temp.in = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{rd.frmt}{Vector of characters. The file format to read. 
By default it will be read  as a  R object using the
\code{'readRDS'} argument, but it can be read as plain text using the
\code{'readTXT'} argument. See details.}

\item{path}{Vector of characters. Path to the input file.}

\item{API.Key}{If the user has a google's API key, it must be used here,
otherwise no more than 2500 elevations could be assigned.

See details.}

\item{starts.in}{Numeric vector. Row number where should start the process.}

\item{round.coord}{Integer. Number of digits to keep in coordinates.}

\item{save.name}{Vector of characters. Name of the output file.}

\item{wrt.frmt}{Vector of characters. Format to save the output
file. By default it will be written  as a  R object using 
\code{'saveRDS'} argument, but it can be saved as plain text using 
\code{'saveTXT'} argument. See details.}

\item{save.assigned.in}{Vector of characters. Path to 
the output file with the coordinates whose elevation was assigned. 
Output file will be called <alt.assigned>.}

\item{save.unassigned.in}{Vector of characters. Path to 
the output file with the coordinates whose elevation was not assigned. 
Output file will be called <alt.unassigned>. See details.}

\item{save.temp.in}{Vector of characters. Path to the temporal file 
as backup of the process. If the process is stopped, the user could reset it 
assigning the row number in \code{start.in} parameter and the process will start again
using temporal file.}
}
\value{
A table data.frame class with coordinates and their respective elevation. 
A table with coordinates whose elevations were not assigned. A table as backup of the process.
}
\description{
Given a coordinates vector, download the elevation data 
from The Google Maps Elevation API [1].
}
\details{
For more details about the formats to read and/or write, see 
\code{\link{readAndWrite}} function.

The coordinates whose elevations were not assigned could be checked with
\code{\link{assignElevation}}. This function can download directly from
Google Maps Elevation API[1]

Additional information at: 
\url{https://developers.google.com/maps/documentation/elevation/#api_key}
 

The headers of the input file must follow the Darwin Core standard [2]. 
The user can see the guide using data('ID_DarwinCore) command.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] The Google Maps Elevation API.  Available online at \url{https://developers.google.com/maps/documentation/elevation/intro}

[2] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{assignElevation}}
}

