% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Europa.R
\docType{data}
\name{Europa}
\alias{Europa}
\title{Vectorial layer: Europan continent}
\format{Format class \code{'SpatialPolygonsDataFrame'} [package \code{'sp'}].}
\source{
\url{http://r-forge.r-project.org/projects/maptools/}
}
\usage{
data(Europa)
}
\description{
Map of the Europan continent. Shapefile of countries borders. This map was
extracted from data(wrld_simpl) from maptools package [1]. The \code{'Europa'} object uses the WGS84
datum.
}
\examples{
\donttest{
data(Europa)
plot(Europa)}
}
\references{
[1] Bivand  et al. (2015) maptools: Tools for Reading and Handling Spatial Objects. CRAN. Version 0.8-37. Available online at \url{http://r-forge.r-project.org/projects/maptools/}
}
\keyword{datasets}

