% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\docType{data}
\name{Example}
\alias{Example}
\title{Input file to example Rmarkdown}
\format{File modified from GBIF, occurrences for 35 species of mammals with distribution on the biogeographic Choco.}
\usage{
data(Example)
}
\description{
Input file to example Rmarkdown [1].
}
\examples{
 \donttest{data(countryNames)}

}
\references{
[1] R-Alarcon V. and Miranda-Esquivel DR. (submitted) geocleaMT: An R
package to cleaning geographical data from electronic biodatabases.
}
\keyword{datasets}

