% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifDownSp.R
\name{gbifDownSp}
\alias{gbifDownSp}
\title{Download occurrences for a list of species}
\usage{
gbifDownSp(sp.name = NULL, taxon.key = NULL, genus = NULL,
  epithet = NULL, starts.in = 1, wrt.frmt = "saveRDS",
  save.download.in = NULL, ...)
}
\arguments{
\item{sp.name}{Vector of characters. A species list to download. See details.}

\item{taxon.key}{Vector of characters. A codes list, these codes are assigned by GBIF [1]
to each species. See details.}

\item{genus}{Vector od Characters. A genera list to download.}

\item{epithet}{Vector of characters. A epithet list to download.}

\item{starts.in}{Numeric vector. Row number to start the process.}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using
\code{'saveRDS'} argument, but it can be saved as plain text using
\code{'saveTXT'} argument. See details.}

\item{save.download.in}{Character.  Vector of characters. Path to
the output file for each species downloaded.}

\item{...}{All parameters that can parsed to \code{\link[rgbif]{occ_search}}
function from \code{rgbif} package [2] as limit, basisOfRecord, country, 
publishingCountry, lastInterpreted, geometry, collectionCode, institutionCode and year.}
}
\value{
Save one file by species. This file is a list of all occurrences recorded from GBIF.
}
\description{
Download occurrences from the Global Biodiversity Information Facility (GBIF) [1]
}
\details{
Additional parameters (...) see \code{\link[rgbif]{occ_search}}
function from \code{rgbif} package.
The user can only use one parameter at a time: \code{sp.name}, \code{taxon.key} or \code{genus}
and/or \code{epithet}.

For  \code{sp.name} parameter, a list of species names  could be assigned.
The species names must have genus and epithet, separated by a single
space.

For \code{taxon.key} parameter the list have to be the ID number that assigns GBIF
to each taxonomical level.


The headers of the input file must follow the Darwin Core standard [3].
The user can see the guide using data('ID_DarwinCore) command.

For more details about the formats to read and/or write, see
\code{\link{readAndWrite}} function.
}
\note{
See:
R-Alarcon V, Miranda-Esquivel DR.(in prep) geocleaMT: An R package to cleaning
geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1]  Global Biodiversity Information Facility. Available online at \url{ http://www.gbif.org/}. 


[2] Chamberlain, S. et al. 2015. rgbif: Interface to the Global Biodiversity Information Facility `API'. R package version 0.9.0. The Comprehensive R Archive Network (CRAN). Available online at \url{http://cran.r-project.org/package=rgbif}.


[3] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}

\code{\link[rgbif]{occ_search}}
}

