% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBISlist.R
\docType{data}
\name{IBISlist}
\alias{IBISlist}
\title{Invasive species list}
\format{A data frame with 65 observations on the following 6 variables.}
\usage{
data(IBISlist)
}
\description{
List of invasive species recorded in Island Biodiversity and Invasive Species (IBIS) [1].
}
\examples{
 \donttest{data(IBISlist)}

}
\references{
[1] Kells, S. S. and Worswick, C. 1997. An introduction to the IBIS database. Melbourne Institute of Applied Economic and Social Research.
}
\keyword{datasets}

