% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID_DarwinCore.R
\docType{data}
\name{ID_DarwinCore}
\alias{ID_DarwinCore}
\title{ID Darwin Core standard}
\format{A data frame with 224 observations on the following 2 variables: ID and HeaderName.}
\usage{
data(ID_DarwinCore)
}
\description{
Headers following the Darwin Core standard [1].
}
\examples{
data(ID_DarwinCore)

}
\references{
[1] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\keyword{datasets}

