% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invasiveSp.R
\name{invasiveSp}
\alias{invasiveSp}
\title{Separate invasive species}
\usage{
invasiveSp(data = NULL, rd.frmt = "readRDS", path = NULL, starts.in = 1,
  save.Sp.list = TRUE, wrt.frmt = "saveRDS", save.foreign.in = NULL,
  save.non.foreign.in = NULL, save.temp.in = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{rd.frmt}{Vector of characters. The file format to read.
By default it will be read  as a  R object using the
\code{'readRDS'} argument, but it can be read as plain text using the
\code{'readTXT'} argument. See details.}

\item{path}{Vector of characters. Path to the input file.}

\item{starts.in}{starts.in Numeric vector. Row number to start the process.}

\item{save.Sp.list}{Logic. If \code{'TRUE'}, two files are saved, one has the
list of invasive species and other has the list of non-invasive species,
both have two columns genus and  specific epithet and it will be used as
input for the \code{\link{gbifDownSp}} function.}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using
\code{'saveRDS'} argument, but it can be saved as plain text using
\code{'saveTXT'} argument. See details.}

\item{save.foreign.in}{Vector of characters. Path to
write the output file with species that are categorized as invasive.}

\item{save.non.foreign.in}{Vector of characters. Path to
write the output file with species that are not categorized as invasive.}

\item{save.temp.in}{Vector of characters. Path to the temporal file
as backup of the process. If the process is stopped, the user could reset it
assigning the row number in \code{start.in} parameter and the process will start again
using the temporal file.}
}
\value{
A table data.frame class with a list of invasive species.
A table data.frame class with a list of non-invasive species.
}
\description{
Separate species categorized as invasive for The Global
Invasive Species Information Network (ISSG) [1] and Island Biodiversity
and Invasive Species (IBIS) [2].
}
\details{
For more details about the formats to read and/or write, see
\code{\link{readAndWrite}} function.

If the argument in the \code{save.Sp.list} parameter is \code{'TRUE'},
the files will be saved in the path assigned in the \code{path} parameter.

The headers of the input file must follow the Darwin Core standard [3].
The user can see the guide using data('ID_DarwinCore) command.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Lowe, S. et al. 2000. 100 of the world's worst invasive alien species: A selection from the global invasive species database. The Invasive Species Specialist Group (ISSG) a specialist group of the Species Survival Commission (SSC) of the World Conservation Union (IUCN).

[2] Kells, S. S. and Worswick, C. 1997. An introduction to the IBIS database.  Melbourne Institute of Applied Economic and Social Research.

[3] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

