% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISO2.R
\docType{data}
\name{ISO2}
\alias{ISO2}
\title{ISO 3166-1 alpha-2}
\format{A vector class Factor with 246 ISO-2 acronyms.}
\usage{
data(ISO2)
}
\description{
List of International Organization for Standardization (ISO 3166-1 alpha-2) [1].
}
\examples{
 \donttest{data(ISO2)}

}
\references{
[1]ISO 2015. International standard for country codes and codes for their subdivisions. Country codes - iso 3166.  International Organization for Standardization. Available online at \url{http://www.iso.org/iso/country_codes}
}
\keyword{datasets}

