% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISO3.R
\docType{data}
\name{ISO3}
\alias{ISO3}
\title{ISO 3166-1 alpha-3}
\format{A vector class Factor with 246 ISO-3 acronyms.}
\usage{
data(ISO3)
}
\description{
List of International Organization for Standardization (ISO 3166-1 alpha-3) [1].
}
\examples{
 \donttest{data(ISO3)}

}
\references{
[1]ISO 2015. International standard for country codes and codes for their subdivisions. Country codes - iso 3166.  International Organization for Standardization. Available online at \url{http://www.iso.org/iso/country_codes}
}
\keyword{datasets}

