% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOToCountry.R
\name{ISOTocountry}
\alias{ISOTocountry}
\title{Convert from  International Organization for Standardization (ISO 3166-1 alpha-2 and  alpha-3)
to country name (English).}
\usage{
ISOTocountry(countryCode = NULL, iso = 2)
}
\arguments{
\item{countryCode}{Data.Frame class. A vector with the ISO-2or ISO-3 Codes.}

\item{iso}{numeric (2 or 3). To define the code whether ISO-2 (iso=2, default value) or ISO-3 (iso=3).}
}
\value{
A table with two columns (countryCode, country)
}
\description{
Given the ISO-2 or ISO-3 code [1], it  will assign the corresponding country name in English.
}
\details{
The user can see the ISO 3166-1 alpha-2 or alpha-3 valid for each country using
the line of code : data(ISO2) or data(ISO3).
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(in prep) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] ISO 2015. International standard for country codes and codes for their subdivisions. Country codes - iso 3166. International Organization for Standardization. Available online at http://www.iso.org/iso/country_codes .
}
\seealso{
\code{\link{countryToCentroid}}
}

