% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanPropinquity.R
\name{meanPropinquity}
\alias{meanPropinquity}
\title{Mean Propinquity}
\usage{
meanPropinquity(coord.table = NULL, calculatedBy = "species",
  wrt.frmt = "saveTXT", save.info.in = NULL, plot.dist = FALSE,
  plot.onMap = FALSE, save.plot.in = NULL)
}
\arguments{
\item{coord.table}{data frame class. Table with coordinates (decimalLongitude,
decimalLatitude). If the process will be calculated by species, the table must
include the species column.}

\item{calculatedBy}{Vector of characters. If the argument is \code{'species'}, 
the mean propinquity will be calculated for each species. If it is \code{'whole'} 
the propinquity will be calculated as a whole.}

\item{wrt.frmt}{Vector of characters. Format to save the output file.
By default it will be written as a R object using 
\code{'saveRDS'} argument, but it can be saved as plain text using 
\code{'saveTXT'} argument. See details.}

\item{save.info.in}{Vector of characters. Path to the output 
file with information about Mean propinquity, median, mode, minimal spanning 
tree; by species or as a whole.}

\item{plot.dist}{Logical vector. If \code{'TRUE'} a density plot for the distance 
between occurrences will be plotted and saved, by species or as a whole. If 
\code{'FALSE'} there will be no plot.}

\item{plot.onMap}{Logical vector. If \code{'TRUE'} a map of occurences  will be 
plotted and saved, by species or as a whole. If \code{'FALSE'} there will be no plot.}

\item{save.plot.in}{Vector of characters. Path to the output 
file for each plot by species or as a whole.}
}
\value{
The output file will be saved with all species concatenated.
}
\description{
Calculate the mean propinquity[1] and some descriptors of the distribution.
}
\details{
If the process uses the parameters plot.dist and plot.onMap as \code{'TRUE'}, the process will be slow.
The headers of the input file must follow the Darwin Core standard [2]. 
The user can see the guide using data('ID_DarwinCore) command.
For  more details about the formats to  write, see 
\code{\link{readAndWrite}} function.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR. (submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Rapoport. E.H. 1975 Areografia:  Estrategias  Geograficas  de  las  Especies. Mexico, Fondo de Cultura Economica

[2] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

