% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsAtSea.R
\name{pointsAtSea}
\alias{pointsAtSea}
\title{Separate occurrences on Earth and at sea}
\usage{
pointsAtSea(data = NULL, path = NULL, rd.frmt = "readRDS",
  wrt.frmt = "saveRDS", save.OnEarth.in = NULL, save.AtSea.in = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{path}{Vector of characters. Path to the input file.}

\item{rd.frmt}{Vector of characters. The file format to read. 
By default it will be read  as a  R object using 
\code{'readRDS'} argument, but it can be read as plain text using 
\code{'readTXT'} argument. See details.}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using 
\code{'saveRDS'} argument, but it can be saved as plain text using 
\code{'saveTXT'} argument. See details.}

\item{save.OnEarth.in}{Vector of characters. Path to the output 
file for each  species with the coordinates on earth.}

\item{save.AtSea.in}{Vector of characters. Path to the output 
file for each  species with the coordinates at sea.}
}
\value{
By each species, if this is the case, two output files 
data.frame class: The first one has the coordinates at sea for the species and
the second one has the coordinates on earth for the same species.
}
\description{
Delete duplicates, separating records
on Earth and at sea for each species.
}
\details{
For  more details about the formats to read and/or write, see 
\code{\link{readAndWrite}} function.

The headers of the input file must follow the Darwin Core standard [1]. 
The user can see the guide using data('ID_DarwinCore) command.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

