% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndWrite.R
\name{readAndWrite}
\alias{readAndWrite}
\title{Read and write tables}
\usage{
readAndWrite(action = NULL, frmt = NULL, path = NULL, name = NULL,
  object = NULL)
}
\arguments{
\item{action}{Vector of characters. If \code{'read'} string, the function will 
read the file. If \code{'write'} string, the function will write output file.}

\item{frmt}{Vector of characters. The format to read or write can be of 
two types: plain text \code{'readTXT'}/\code{'saveTXT'} or a R object
\code{'saveRDS'}/\code{'saveRDS'}.}

\item{path}{Vector of characters. Path to the input file or the 
destination Path to the output file.See details.}

\item{name}{Vector of characters. Name of the input file or name of file to
write.}

\item{object}{Vector of characters. Object to write. See details.}
}
\value{
For action = \code{'read'}, load the file given for the user. 
For action = \code{'write'}, save an output file. Both depend on the frmt parameter.
}
\description{
Read or write tables in two specific formats for standardizing the 
cleaning process.  This function reads the input and writes the output.
}
\details{
It reads and/or writes a file in plain text format (txt) using the 
functions \code{\link{read.table}} and/or \code{\link{write.table}}. 
To maintain the standard manipulation the arguments for both 
functions \code{\link{read.table}} and \code{\link{write.table}} parameters, are the default,
except for the parameters: sep= 'TAB', header= T, quotes= F and rownames = F.

If the argument in the action parameter is \code{'write'}, the argument in object 
and name parameters must be assigned. But, if the argument in 
the action parameter is \code{'read'}, the argument in object  parameter
must be omitted.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(in prep) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\seealso{
\code{\link{saveRDS}}

\code{\link{write.table}}
}

