% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitGeoref.R
\name{splitGeoref}
\alias{splitGeoref}
\title{Separate geo-referenced records}
\usage{
splitGeoref(data = NULL, rd.frmt = "readRDS", path = NULL, min.occ = 3,
  round.coord = 4, wrt.frmt = "saveRDS", save.min.occ.in = NULL,
  save.georef.in = NULL, save.ungeoref.in = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{rd.frmt}{Vector of characters. The file format to read. 
By default it will be read  as a  R object using 
\code{'readRDS'} argument, but it can be read as plain text using 
\code{'readTXT'} argument. See details.}

\item{path}{Vector of characters. Path to the data file(s).}

\item{min.occ}{Numeric vector. Minimal number of georeferenced occurrences 
for that a species can be considered. If the argument is 3, only species 
with 3 or more occurrences will be considered in the process.}

\item{round.coord}{Integer. Number of digits to keep in coordinates.}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using 
\code{'saveRDS'} argument, but it can be saved as plain text using 
\code{'saveTXT'} argument. See details.}

\item{save.min.occ.in}{Character. Path to save the output file of species 
that do not fulfill the minimum required number of occurrences in the parameter 
\code{min.occ}. Default format is  \code{'.RDS'}. Se details.}

\item{save.georef.in}{Vector of characters. Path to write the output 
file for each species with georeferenced records.}

\item{save.ungeoref.in}{Character. Path to save ungeoreferenced data,
a file for each species. Default format is  \code{'.RDS'}. Se details.}
}
\value{
By each species, if this is the case, two output files 
data.frame class: The first one  have the georeferenced and
the second one have the ungeoreferenced
}
\description{
Get only georeferenced records from a database
downloaded from GBIF[1].
}
\details{
For more details about the formats to read and/or write, see 
\code{\link{readAndWrite}} function.

The headers of the input file must follow the Darwin Core standard [2]. 
The user can see the guide using data('ID_DarwinCore) command.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Global Biodiversity Information Facility. Available  online at \url{ http://www.gbif.org/}.

[2] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

