% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spOutPoly.R
\name{spOutPoly}
\alias{spOutPoly}
\title{Obtain species outside of a polygon}
\usage{
spOutPoly(data = NULL, rd.frmt = "readRDS", path = NULL,
  shp.poly = NULL, max.per.out = 5, max.occ.out = 3, execute = T,
  B1 = F, B2 = T, wrt.frmt = "saveRDS", save.inside.in = NULL,
  save.outside.in = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{rd.frmt}{Vector of characters. The file format to read. 
By default it will be read  as a  R object using 
\code{'readRDS'} argument, but it can be read as plain text using 
\code{'readTXT'} argument. See details.}

\item{path}{Vector of characters. Path to the input file.}

\item{shp.poly}{SpatialPolygonDataframe Class. Polygon of the referenced areas.
The area where the species has to be distributed.}

\item{max.per.out}{Numeric vector. Maximum percentege of occurrences that can 
be outside of polygon for the species to be considered as species with 
restricted distribution. See details.}

\item{max.occ.out}{Numeric vector. Absolute maximum number of occurrences that can 
be outside of a polygon for the species to be considered as species with 
restricted distribution. See details.}

\item{execute}{Logical vector. If \code{'TRUE'}, the process will be executed on species.
If \code{'FALSE'} the species will only be classified and an information table will 
be saved.}

\item{B1}{Logical vector. If \code{'TRUE'}, the condition B1T will be applied.
If \code{'FALSE'} the condition B1F will be applied. See details.}

\item{B2}{Logical vector. If \code{'TRUE'}, the condition B2T will be applied.
If \code{'FALSE'} the condition B2F will be applied. See details.}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using}

\item{save.inside.in}{Vector of characters. Path to the output 
file for each  species whose distribution was considered restricted to polygon.}

\item{save.outside.in}{Vector of characters.  Path to the output 
file for each  species whose distribution was considered widespread.}
}
\value{
If execute is \code{'TRUE'} a table data.frame class with information about 
how was the species classified and an output file by species with occurrences. 
If execute is FALSE, a table data.frame class will be saved,  with information about 
how was the species classified.
}
\description{
This function separates species with restricted 
distribution to a polygon or a defined area, following a series
of conditions.
}
\details{
#'
To see the descriptions and details about conditions and parameters the user 
should check  R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.

The condition A, B3 and  C for species classification are executed by
default, and these can not be modified.

The headers of the input file must follow the Darwin Core standard [1]. 
The user can see the guide using data('ID_DarwinCore) command.
For  more details about the formats to read and/or write, see 
\code{\link{readAndWrite}} function.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

