% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackSp.R
\name{stackSp}
\alias{stackSp}
\title{Join species files}
\usage{
stackSp(data = NULL, rd.frmt = "readRDS", path = NULL, save.name = NULL,
  save.staking.in = NULL, wrt.frmt = "saveRDS")
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{rd.frmt}{Vector of characters. The file format to read. 
By default it will be read  as a  R object using the
\code{'readRDS'} argument, but it can be read as plain text using the
\code{'readTXT'} argument. See details.}

\item{path}{Vector of characters. Path to the input file.}

\item{save.name}{Vector of characters. Name of the output file.}

\item{save.staking.in}{Vector of characters. Path to the output 
file.}

\item{wrt.frmt}{Vector of characters. Format to save output
file. By default it will be written  as a  R object using 
\code{'saveRDS'} argument, but it can be saved as plain text using 
\code{'saveTXT'} argument. See details.}
}
\value{
The output file will be saved with all species concatenated.
}
\description{
Join all independent files into a single multi-species file.
}
\details{
The headers of the input file must follow the Darwin Core standard [1]. 
The user can see the guide using data('ID_DarwinCore) command.
For  more details about the formats to read and/or write, see 
\code{\link{readAndWrite}} function.
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

