% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefulSp.R
\name{usefulSp}
\alias{usefulSp}
\title{Useful species}
\usage{
usefulSp(data = NULL, path = NULL, cut.off = 3, rd.frmt = "readRDS",
  wrt.frmt = "saveRDS", save.useful.in = NULL, save.useless.in = NULL)
}
\arguments{
\item{data}{Vector of characters. Name of the input file.}

\item{path}{The path for the file or files.}

\item{cut.off}{Numeric. Minimal (absolute) number of occurrences for a given 
species, the default value is 3, as the minimal number of points to draw a polygon.}

\item{rd.frmt}{Character. Input file format, by default it uses a .RDS format. 
See details.}

\item{wrt.frmt}{Character. Output file format,  by default it uses a \code{'.RDS'}
format, although it could save a txt table. See details.}

\item{save.useful.in}{String. Output file path for useful species (those with 
occurrences larger than the \code{cut.off} value. See details.}

\item{save.useless.in}{String. Output file path for useless species (those with 
occurrences smaller than the \code{cut.off} value. See details.}
}
\value{
A data.frame object and a file for each species.
}
\description{
Get  species with a minimum number of occurrences.
}
\details{
To check input/output formats see \code{\link{readAndWrite}}.
Column names must follow Darwin Core2[1] standard. See data('ID_DarwinCore).
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(submitted) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] Wieczorek, J. et al. 2012. Darwin core: An evolving community-developed biodiversity data standard. 
PloS One 7: e29715.
}
\seealso{
\code{\link{readAndWrite}}
}

