% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Africa.R
\docType{data}
\name{Africa}
\alias{Africa}
\title{Vectorial layer: African continent}
\format{Format class \code{'SpatialPolygonsDataFrame'} [package \code{'sp'}][1].}
\source{
\url{http://r-forge.r-project.org/projects/maptools/}
}
\usage{
data(Africa)
}
\description{
Map of the African continent. Shapefile of countries borders. This map was
extracted from data(wrld_simpl) from maptools package. The \code{'Africa'} object uses the WGS84
datum.
}
\examples{
\donttest{
data(Africa)
plot(Africa)}
}
\references{
[1] Bivand et al. (2015) Maptools: Tools for Reading and Handling Spatial Objects. CRAN. Version 0.8-37. Available online at \url{http://r-forge.r-project.org/projects/maptools/}
}
\keyword{datasets}

