% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/America.R
\docType{data}
\name{America}
\alias{America}
\title{Vectorial layer: American continent}
\format{Format class \code{'SpatialPolygonsDataFrame'} [package \code{'sp'} [2]].}
\source{
\url{http://r-forge.r-project.org/projects/maptools/}
}
\usage{
data(America)
}
\description{
Map of the American continent. Shapefile of countries borders. This map was
extracted from data(wrld_simpl) from maptools package [1]. The \code{'America'} object uses the WGS84
datum.
}
\examples{
\donttest{
data(America)
plot(America)}
}
\references{
[1] Bivand  et al. (2015) maptools: Tools for Reading and Handling Spatial Objects. CRAN. Version 0.8-37. Available online at \url{http://r-forge.r-project.org/projects/maptools/}
[2] Pebesma et al. (2015) sp: Classes and methods for spatial data.  CRAN. Version 1.2-2. Available online at \url{https://cran.r-project.org/web/packages/sp/}
}
\keyword{datasets}

