% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOToCountry.R
\name{ISOTocountry}
\alias{ISOTocountry}
\title{Convert from the International Organization for Standardization (ISO 3166-1 alpha-2 or alpha-3) code
to country name (English).}
\usage{
ISOTocountry(countryCode = NULL)
}
\arguments{
\item{countryCode}{Data.Frame class. A vector with the ISO-2or ISO-3 Codes.}
}
\value{
A table with two columns (countryCode, country)
}
\description{
Given the lenght of the ISO country code provided (2=ISO-2 or 3=ISO-3[1]), the function will assign the corresponding country name in English.
}
\details{
The user can check the ISO 3166-1 alpha-2 or alpha-3 code, valid for each country using
the line of code : data(ISO2) or data(ISO3).
}
\note{
See:
R-Alarcon V. and Miranda-Esquivel DR.(in prep) geocleaMT: An R package to
cleaning geographical data from electronic biodatabases.
}
\author{
R-Alarcon Viviana and Miranda-Esquivel Daniel R.
}
\references{
[1] ISO 2015. International standard for country codes and codes for their subdivisions. Country codes - iso 3166. International Organization for Standardization. Available online at http://www.iso.org/iso/country_codes .
}
\seealso{
\code{\link{countryToCentroid}}
}

